requirements.txt
numpy
pandas
ipykernel
kagglehub
tensorflow
keras
pillow
scikit-learn
seaborn
streamlit

-e .